Organized code in a cleaner way. Also, everytime someone signs up for an event their email address gets inserted into firestore. If the number of signups is greater than max attendees then it will not send email.
